editados datos en las entidades

+++ app/src/main/java/com/example/taller4/database/entities/Author.kt
@@ -14,5 +14,5 @@ data class Author (
     val publisher_last_name:String
 ){
     @PrimaryKey(autoGenerate = true)
-    var id:Long=0
+    var id:int=0
 }
